samplegen rename mode
