+++ src/MAUI/Maui.Samples/Samples/Layers/ConfigureElectronicNavigationalCharts/ConfigureElectronicNavigationalCharts.xaml
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage x:Class="ArcGIS.Samples.ChangeEncDisplaySettings.ChangeEncDisplaySettings"
+<ContentPage x:Class="ArcGIS.Samples.ConfigureElectronicNavigationalCharts.ConfigureElectronicNavigationalCharts"
              xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:esriUI="clr-namespace:Esri.ArcGISRuntime.Maui;assembly=Esri.ArcGISRuntime.Maui">
