+++ src/WinUI/ArcGIS.WinUI.Viewer/Samples/Layers/ConfigureElectronicNavigationalCharts/ConfigureElectronicNavigationalCharts.jpg
