+++ src/WinUI/ArcGIS.WinUI.Viewer/Samples/Layers/ConfigureElectronicNavigationalCharts/ConfigureElectronicNavigationalCharts.xaml
@@ -1,17 +1,21 @@
-﻿<UserControl x:Class="ArcGIS.WPF.Samples.ChangeEncDisplaySettings.ChangeEncDisplaySettings"
+﻿<UserControl x:Class="ArcGIS.WinUI.Samples.ConfigureElectronicNavigationalCharts.ConfigureElectronicNavigationalCharts"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013">
+             xmlns:esriUI="using:Esri.ArcGISRuntime.UI.Controls">
     <UserControl.Resources>
+        <Style TargetType="TextBlock">
+            <Setter Property="FontWeight" Value="SemiBold" />
+            <Setter Property="Margin" Value="5,15,0,0" />
+        </Style>
         <Style TargetType="RadioButton">
-            <Setter Property="Margin" Value="0,5,5,10" />
+            <Setter Property="Margin" Value="5,5,0,0" />
         </Style>
     </UserControl.Resources>
     <Grid>
-        <esri:MapView x:Name="MyMapView" />
-        <Border Width="225" Style="{StaticResource BorderStyle}">
+        <esriUI:MapView x:Name="MyMapView" />
+        <Border Style="{StaticResource BorderStyle}">
             <StackPanel>
-                <TextBlock FontWeight="Bold" Text="Color scheme" />
+                <TextBlock Text="Color scheme" />
                 <StackPanel Orientation="Horizontal">
                     <RadioButton x:Name="DayRadioButton"
                                  Content="Day"
@@ -24,7 +28,7 @@
                                  Content="Night"
                                  GroupName="ColorScheme" />
                 </StackPanel>
-                <TextBlock FontWeight="Bold" Text="Area symbolization type" />
+                <TextBlock Text="Area symbolization type" />
                 <StackPanel Orientation="Horizontal">
                     <RadioButton x:Name="PlainAreaRadioButton"
                                  Content="Plain"
@@ -34,9 +38,9 @@
                                  Content="Symbolized"
                                  GroupName="Area" />
                 </StackPanel>
-                <TextBlock FontWeight="Bold" Text="Point symbolization type" />
+                <TextBlock Text="Point symbolization type" />
                 <StackPanel Orientation="Horizontal">
-                    <RadioButton x:Name="PaperchartRadioButton"
+                    <RadioButton x:Name="PaperPointRadioButton"
                                  Content="Paper Chart"
                                  GroupName="Point"
                                  IsChecked="True" />
