+++ src/WinUI/ArcGIS.WinUI.Viewer/Samples/Layers/ConfigureElectronicNavigationalCharts/readme.md
@@ -1,8 +1,8 @@
-# Change ENC display settings
+# Configure electronic navigational charts
 
 Configure the display of ENC content.
 
-![Image of change ENC display settings app](changeencdisplaysettings.jpg)
+![Image of change ENC display settings app](ConfigureElectronicNavigationalCharts.jpg)
 
 ## Use case
